Wrote allocate_timestamps.

+++ test/test_Preprocessor.py
@@ -89,24 +89,20 @@ class TestPreprocessor(unittest.TestCase):
         assert_frame_equal(correct, dates)
 
     def test_allocates_timestamps(self):
-        # TODO Write that method!
         # Make a shortcut to the real, raw typical csv file, pretending it's clean
-        os.symlink(self.preprocessor.raw_csv_dir + '/shcu_typical_data.csv', self.preprocessor.data_dir + '/fake_clean_data.csv')
-        self.preprocessor.allocate_timestamps('fake_clean_data.csv',
-                                              [0.6, 0.2, 0.2],
-                                              ['typical_training_timestamps',
-                                               'typical_validation_timestamps',
-                                               'typical_testing_timestamps'])
+        if not os.path.exists(self.preprocessor.data_dir + '/fake_clean_data.csv'):
+            os.symlink(self.preprocessor.raw_csv_dir + '/shcu_typical_data.csv', self.preprocessor.data_dir + '/fake_clean_data.csv')
+        filenames = ['typical_training_timestamps', 'typical_validation_timestamps', 'typical_testing_timestamps']
+        self.preprocessor.allocate_timestamps('fake_clean_data.csv', [0.6, 0.2, 0.2], filenames)
         train = int(check_output(['wc', self.preprocessor.data_dir + '/typical_training_timestamps']).split()[0])
         valid = int(check_output(['wc', self.preprocessor.data_dir + '/typical_validation_timestamps']).split()[0])
         test = int(check_output(['wc', self.preprocessor.data_dir + '/typical_testing_timestamps']).split()[0])
         total = train + valid + test
-        correct_total = int(check_output(['wc', self.preprocessor.raw_csv_dir + '/fake_clean_data.csv']).split()[0]) - 1
+        correct_total = int(check_output(['wc', self.preprocessor.data_dir + '/fake_clean_data.csv']).split()[0]) - 1
         self.assertEqual(correct_total, total)
         self.assertTrue(0.5 < train / total < 0.7)
         self.assertTrue(0.1 < valid / total < 0.3)
         self.assertTrue(0.1 < test / total < 0.3)
-
 
 
 if __name__ == '__main__':
